added octosonar_bench.ino
This is example is what I use to drive the bench test for new boards.

The jig is made of two V2.1 boards that had bad silkscreen and
pogo-pins. The lower board has a PCF8574 at 0x20 so it can spy on the
TRIG signals being sent to the test board. The echo pins are driven from
arduino digital pins.

I ran into horrible problems writing things to the test board and
consistently seeing the result when read back from the jig. At first I
thought it might be a pullup resistor issue, but finally I isolated it
to some kind of timing issue and/or interaction with Serial or some othe
rinternal arduino interrupts.

So with all the debug chatter commented out, it gets all the way through
and gives me a green. Actual boards mounted in actual projects seem to
work fine, so this may just be some wierd interaction specific to the
test jig.