Add working support for OctoSonarX2
The ifdef settings in the previous version did not actually work in
practice. The two board types are now derived classes from a base, and
the code alters behavior when it has to write to the 16 bit PCF8575 on
the X2.  The only api difference is that the active value is now
uint16_t not uint8_t. This should be backwards compatible.