Update octosonarX2.brd to 1.0.1
Moved the 5V line away from the screw hole by moving it to bottom layer. This never caused any reported problem, but I was worried about it and was supplying a plastic screw with each unit just in case. This error was introduced between v0 and V1 when I re-did all the power lines at 10mil.